Revert "Fix access method creation to use AO am handlers correctly"
This reverts gpdb-postgres-merge commit 9809bcae17377170b6e5e97faf33c6315a3a12be.

The functionality to identify based on the handler oid is too esoteric
to justify the cost of extra maintenance overhead. Using the burned in
table AM oid for AO and AOCO should be sufficient.

Conflicts:

* We keep the create_am_gp.sql test as the cases add valuable
coverage but we adapt it to use the built-in access methods.