Resolve a FIXME in prepare_AlterTableStmt_for_dispatch()
GPDB has specific ALTER TABLE commands for partition tables, which
are handled differently than other AT commands: instead of being
executed on QD first and then dispatched to QEs after all the other
commands are done, they are dispatched immediately in ATExecCmd()
and executed on QEs at the same time as QD. This results in the
possibility that AT commands could be executed in different order
in QEs than in QD.

The FIXME comment suggests to make those AT partition commands
same as other AT commands, but that could be very invasive, e.g.:

1. The AT partition commands generate a lot of subcommands that
are not in the AT family (see ATExecGPPartCmds()). We cannot put them
in the AT workqueue as AT subcommands. At minimal we would need
separate code to dispatch them.
2. Many of the subcommands implicitly dispatch to QEs (e.g.
CreateStmt/DropStmt). For those, we need to have workarounds to
to execute them w/o dispatching.
3. Some subcommand needs partial results from other subcommands
before executing (e.g. AT_PartSplit).

Therefore, we take a less invasive approach which is to group
the GP-specific commands together and execute them before all the
other commands. Since the other commands' order is the same on QD
and QEs, we resolved the original issue.