Remove gp_allow_rename_relation_without_lock
This GUC was introduced in de4e95a6a914 in order to workaround
a specific case where we couldn't rename a relation if its metadata
is corrupted. When this GUC is ON, we don't really open the relation
which would throw error with the aforementioned metadata corruption.

Now removing this GUC for 7X because its benefit is questionable:
(1) It seems to be a one-time workaround trick, and has been obsolete
for a long time. There's no Slack/JIRA/Github history that could show
it has been actively used by support);
(2) There is better alternative to workaround the issue, which is to
simply fix the metadata. Leaving a hole like this in the product does
not feel safe. It creates needless divergence from the upstream too.