Do not update old reloptions during ALTER TABLE in the case of table rewrite
When setting reloptions, we check if there's any change in the AO/CO
options. If so, we will perform a table rewrite, which involves
copying data from the existing table to a new table. In that case,
we should not update pg_class.reloptions like what we previously do.
That is because copying from the existing table might have problem
when its reloptions do not match with its storage format. So, now we
will pass the new reloptions over through the AlteredTableInfo struct.

This is essentially a code refactor. No actual behavior change. But
it avoids the inconsistent interim state of the table which I think
is the right thing to do.