Using __gp_aoseg/__gp_aocsseg in missing/orphaned file views
The __gp_aoseg/__gp_aocsseg functions provide more details such as the eof
of segments. Use them for the check missing/orphaned file views, and make two
changes:

* For checking missing files, ignore those w/ eof<=0. They might be recycled
but their aoseg/aocsseg entries are still there.

* For checking orphaned files, ignore those that still have base file (w/o any
extension number) being present. Those might be the ones that have been
truncated  but not yet removed. They might be the ones that are left behind
when column is rewritten during ALTER COLUMN. Now the checking logic becomes:
only if the base file is orphaned too, we will report all the extensions along
with it.