Disable autovacuum for flaky test gp_check_files
After #16469 we are still having the issue which however seems to be associated
with a few platforms. Now doing the original fix which is to disable autovacuum
for the test to make sure it won't get in the way of gp_check_orphaned_files in
any way. Meanwhile will look at the platform-specific issue separately.