Do not create schema gp_toolkit within extension gp_toolkit
We did that when we made gp_toolkit an extension. All objects in the extension
are under 'gp_toolkit' schema, but the extension itself is created under
'public' schema (pg_extension reflects that and there's a pg_depend linkage).
However, conventionally 'public' schema should not contain anything other than
user themselves created. And in many cases the user would opt to drop the
'public' schema - and they now have to drop gp_toolkit extension before doing
that, which is not nice.

So now we do not create the schema within the extension, but do that via the
control file which implicitly creates the schema and use it for the extension.
This removes the dependency on 'public' and also conforms the conventions
better. Because we cannot remove "CREATE SCHEMA gp_toolkit" from gp_toolkit--1.0.sql,
create a new default extension version 1.3 to achieve that.

This fixes it for new GPDB installations but existing GPDB7 installations will
continue to have gp_toolkit registered under the 'public' schema which is a
decision based on both impact (not a bug but rather a nice-to-have feature) and
the state of 7X (not many existing installaions). If they re-create gp_toolkit
extension after upgrading to a version containing this change, everything would
be aligned with new installations.

Restore a regress test case that drops the 'public' schema w/o dropping
anything else.

P.S. we continue to explicitly specify the schema name 'gp_toolkit.XXX' in the
extension script, just for more clarity as many views/functions use other
views/functions internally and need to have the right schema to look at things.