Directly use gp_toolkit.__check_orphaned_files in test
The wrapper view for checking orphaned files gp_toolkit.gp_check_orphaned_files
has a bunch of strict activity checks for robustness of the results. However,
in dev environment they had caused the test to be quite flaky due to background
activities. To address them one-by-one is too costly. I think it is OK to skip
the checks in test and directly use the underlying views to check orphaned
files. With that, we can also remove past attempts to address the flakiness.