Fix rownum comparing logic in AppendOnlyExecutorReadBlock_BindingInit
When the row we are checking equals to the lastrownum of an attribute, that
means the row is missing that attribute. As a result, we should stop the loop
because we've found the first attribute that's missing in the row. Therefore,
the logic here is flawed and we should really be doing ">" instead of ">=" here.

However, interestingly, the bug couldn't be manifested, because we won't have a
real row with a row number that's equal to the lastrownum in
pg_attribute_encoding. This is because whenever we are inserting a row like
that, we'll always request more fastsequence numbers (logic is in
appendonly_insert). So if we add column at this point, the lastrownum for the
new column will point to some rownum that's not a real row. Example:

	postgres=# create table ao(a int) using ao_row distributed replicated;
	CREATE TABLE
	postgres=# insert into ao select * from generate_series(1,100);
	INSERT 0 100
	postgres=# alter table ao add column b int;
	ALTER TABLE

	-- on one of the segment, lastrownums is not 100 but 200 because that's what the INSERT had requested.
	postgres=# select * from pg_attribute_encoding where attrelid::regclass::text = 'ao';
	 attrelid | attnum | filenum | lastrownums | attoptions
	----------+--------+---------+-------------+------------
	   272368 |      2 |       1 | {0,200}     |
	(1 row)