Do not write missing-mode column data during ADD COLUMN
(separate commit for now to make it easier for reviewer who reviewed the earlier
approach, will be squashed with the commit "Support missing-mode ADD COLUMN for AOCO tables")

Here's an updated approach to support missing-mode column: we do not write any
table data at all when adding missing-mode column. As a result, scanning the
missing-mode column alone would not get the correct row numbers, instead we
levarage another complete column (i.e. not missing) if it appears in the
projection column list. If there is no complete column in the projection list,
we pick the smallest complete column in the table to scan. In rare situation
the user tries to drop the last complete column in the table, we will rewrite
another column to make it complete and throw a warning.

About block directory: the new column will not have block directory entry even
there is index on the table. If there's a need to find the block directory entry
for the missing column, we will act accordingly: (1) we first check whether the
missing value is supposed to exist in the column or not, by checking the block
directory of the complete column; (2) if it exists, we will fetch the missing
value; otherwise, we will ignore that row.

Kudos to Ashwin Agrawal for the idea.