Support read-committed dtx isolation for hot standby
The previous few commits have removed some road blocks for supporting it. This
commit mainly just deals two more aspects wrt to distributed transactions:

* Initialize latestCompletedGxid during StartupXLOG, and update it while the
  standby replays new transactions.
* Construct an in-progress dtx array when creating distributed snapshot
  according to the shmCommittedGxidArray[] we already keep in the standby.

It was pondered whether or not to add a new WAL type XLOG_RUNNING_DISTRIBUTED_XACTS
similar to XLOG_RUNNING_XACTS. But it seems unnecessary at the moment: we already
have the running dtx information in the checkpoint record. The other information
in the XLOG_RUNNING_XACTS record does not seem to be needed to support
read-committed isolation. There are a few other callers of ProcArrayApplyRecoveryInfo()
that relies on the XLOG_RUNNING_XACTS, but it doesn't seem we have a need to
emulate them for dtx.