Project the anchor column for count(*) for CO tables
Currently for count(*) (or other things that have targetlist and qual as NULL),
we always pick the first column to scan. It is beneficial however to use the
same logic where we pick the anchor column: it is guaranteed to be the smallest
and complete. So use that instead.

Note that: even w/o this patch the first column should always be complete, so
there is no correctness issue, but just performance benefit.