Support repeatable-read dtx isolation for hot standby
In previous commits we've supported hot standby dispatch and read-committed
isolation. In order to support repeatable-read isolation, the only real
complication is just to support the BEGIN...END block. The snapshot selection
and usage for repeatable-read on a hot standby is exactly the same as a primary.

And, the main difference between a single-statement transaction and a
BEGIN...END block is just the DTX context of the QEs: in the former case the
QEs are DTX_CONTEXT_QE_AUTO_COMMIT_IMPLICIT, but in the latter case they
are DTX_CONTEXT_QE_TWO_PHASE_EXPLICIT_WRITER (see setupQEDtxContext()).

We had Assert/ERROR in the code to assume that for EXPLICIT_WRITER, there's
always a valid distributed xid for that transaction. However, that is not the
case for hot standby: a standby never allocates an xid and there's definitely
no use of an xid in its BEGIN...END block. Therefore, all we need to do is just
make sure to not apply this assumption to hot standby. After that, supporting
repeatable-read is a no-op.