Use a temporary gpdb binary in binary swap test
As expected, the optimization of ADD COLUMN for CO tables https://github.com/greenplum-db/gpdb/pull/16765
breaks the binary swap test because it is not forward-compatible (i.e. data
generated with new version doesn't work with old version). We deem it an
exception at the early stage of 7X adoption. In order to fix the binswap
test, we need to use a newer gpdb binary as the "previous-gpdb" version in the
binswap test. For now, we use a temporary gpdb binary that's generated from a
dev pipeline with the latest main branch.

When 7.2.0 is released, we should go back to the old method which is to use a
product release version in the binswap test. So we should revert this commit
and change pivotal-gpdb-pivnet-product-version to be 7.2.0. We have internal
story to track this change.