Support snapshot conflcit on hot standby for ao/co tables and index
The snapshot conflict on hot standby requires the VACUUM running on primary node
to emit the latestRemovedXid during vacuum. For ao/co tables, however, they do
not have xid information stored in them. This xid that is functionally
equivalent to latestRemovedXid of ao/co vacuum, is the the xmin of the
corresponding "AWAITING_DROP" row in the aoseg table. So just like primary node
checks this value to make sure it is safe to truncate the segfile, we now emit
this value in WAL so that hot standby can detect any conflcit. We overload the
existing XLOG_HEAP2_CLEANUP_INFO record as it is not really necessary to
introduce a new WAL type just for ao/co.

For ao/co index, the above latestRemovedXid is also equivalent to the same for
the index tuples we are going to vacuum. The key here is that we are truncating
ao segfile/vacuuming ao index tuples all in the batch of segfiles. So a single
latestRemovedXid can be used for all of them.

In practice, the conflict might be triggered in other places during the ao/co
VACUUM path, most notably during scanning heap which might emits
XLOG_HEAP2_CLEAN records. But our new XLOG_HEAP2_CLEANUP_INFO record acts as
the final guard before conflicting standby query could see wrong results due to
index vacuum & segfile truncation.